Change 6:
BOTH: Transport Mode - currently, a user can specify (select), more than 1 mode of transportation but this is not what we want. User should only be able to select one mode of transport, and then select the driver and the car. Then we need an 'add transport' button, which when clicked will populate all three fields again (Transport Mode | Driver | Vehicle). We need also a fourth drop down in this section called 'Luggage Vehicle' with 3 options: Luggage Car | Luggage Van | Luggage Truck. This field only needs to appear in the first section - not populated when a user selects add transportation. Need to be able to add a total of 5 'transport mode'.

Did some changes, but to clear the air, i think i should discuss this point in detail with client.