Change 7:
BOTH: in the arrivals section, we need a checkbox that says "luggage vehicle". When we pull a report, this box needs to say "no" if not selected, and "yes" if it was selected.

Change Has been Added to the Add Reservation of FSFT, but Yet to be Impleted on the Reports and in Add Reservation of Reservations Also.