Change 7:
BOTH: in the arrivals section, we need a checkbox that says "luggage vehicle". When we pull a report, this box needs to say "no" if not selected, and "yes" if it was selected.

Change 7 Fully Implemented. Both Reports and and Add Reservations Pages Updated.
However its not yet fully Tested.