Change 11:
RESERVATIONS - Add Reservation. Currently, the checkbox for Fast Track is at the top of the page, tied in with the full reservation. This however needs to change - please move this checkbox to the Arrival section. If it is selected for the main arrival, then when we pull a report, it needs to show that the FSFT service is for that specific arrival. If the user needs to "Add Arrival" the checkbox for FSFT needs to also appear in that section. Basically - for every arrival movement, we need to be able to specify whether the guest needs FSFT.

Change Currently Has been Applied to the Reservations as asked. Havent yet thorally tested.