Change 13:
RESERVATIONS: we have a checkbox that says "Add Requirements" and then an option in the dropdown that says Pre booked excursion". If the user selects that option, we need the following fields to populate: Excursion Name, Excursion Date, Pickup Time, Confirmed By Whom, Date of Confirmation, Number of Guests

-=-=-=-=-=
Currently Working on this Change.
Added the Boxes to the Reservations to be Added to the sql.
Also i needed to change the database table added more fields to the fll_arrivals.