Change 8:
BOTH: Hotel Room Number. We need to be able to "add a room" because many reservations have more than one room number. Also when we 'add a room' there needs to be a field for Last Name, so that we know which guest is in that room. Need to be able to "add a room" 5 times.

=-=-=-=-=-==-=
Currently Working on Change 8, However its not yet completed. Only Some design fixtures in Reservations Implemented.