Change 8:
BOTH: Hotel Room Number. We need to be able to "add a room" because many reservations have more than one room number. Also when we 'add a room' there needs to be a field for Last Name, so that we know which guest is in that room. Need to be able to "add a room" 5 times.

=-=-=-=-=-==-=
Currently Working on Change 8, However its not yet completed. Only Some design fixtures in Reservations Implemented.
Javascript Seems to be fixed.
Now only need to Get the Posted Values and Add those values to DB.
9-9-9-9-9-9
Did set up some variables to which posted variables will be saved. However still Need to set up the names of the rooms to rest of arrivals as well.
=-=-=-=-=
Seems to be Completed on Add Reservations Page.

=-=-=-=-=-=-=
Now Currently Working on the FastTrack. Implementing the Same Changes to the Fast Track also.

=-=-=-=-=-=--=
Finally Fully Implemented on Fast Track Also and seems to be that this change has been completed on fll.