Change 6: BOTH: Transport Mode - currently, a user can specify (select), more than 1 mode of transportation but this is not what we want. User should only be able to select one mode of transport, and then select the driver and the car. Then we need an 'add transport' button, which when clicked will populate all three fields again (Transport Mode | Driver | Vehicle). We need also a fourth drop down in this section called 'Luggage Vehicle' with 3 options: Luggage Car | Luggage Van | Luggage Truck. This field only needs to appear in the first section - not populated when a user selects add transportation. Need to be able to add a total of 5 'transport mode'.
Started Work on Change 6,
Currently Working Under Reservations - Add Reservation and in Arrival Section.
Now im fixing only the front end. once the front end is done. will move the the backend saving of data.

=-=-=-=-=-=-=
Front end Seems to be finished, and also did added code for entering data in to the db. plus created a table in db for this change.

But new additions still remaining to be worked on both design and code.