There was a bug in "Assign Airport Reps" to which client had responded with.
I am still having an issue with Point 2
I  am trying to assign an airport rep to a current reservation and it doesn't seem to be working

That Bug has been fixed and a little changes has been implemented.

There was a bug in query. table alias was giving case sensitive error on live server. so that has also been fixed.